requirements/requirements-optionals.txt
# Optional packages which may be used with REST framework.

django-environ==0.4.5  # 12factor inspired environment variables - https://github.com/joke2k/django-environ
psycopg2==2.7.6.1 # Postgres Database Adapter - https://github.com/psycopg/psycopg2
markdown==2.6.11
django-guardian==1.4.9
django-filter==1.1.0
coreschema==0.0.4
pyyaml
ddtrace # Datadog APM integration - http://pypi.datadoghq.com/trace/docs/web_integrations.html#django
requests==2.19.1
uwsgi
